#0024738: user defined Logo superimposed on the muvee [Top-level ticket]
To find out by how much I need to translate the logo image, I multiplied the normalized coordinates with 1.33 or 1.77 depending on the output aspect ratio. 

This creates around errors. I now multiply with 1.77777777 and 1.33333333 depending on the aspect ratio.

Also to prevent z-fighting, I used to add a z-offset of 0.05. It has not been reduced to 0.01. We need to keep the logo and the UGC as close as possible to ensure that there is a high fidelity between the position that the user placed the logo and the actual position in the muvee.