#0024738: user defined Logo superimposed on the muvee [Top-level ticket]
The app leaves a margin of (logo width / 2.0 ) and (logo height / 2.0 ) in its coordinate system.
Stylebase used to check that the bounds of midX and midY are in the range of 0.0 and 1.0.
What Stylebase now does is do the actual subtraction of (logo width / 2.0 ) and (logo height / 2.0 ) to follow the position on the screen more reliably.

Also, z-offset of logo has been increased to 0.03 from 0.01