0016490: Constructor bugs - The Collectors Edition
There is a hat function used to spread out media selection
when possible. This hat function is unfortunately interfering
with selection of highlights, particularly when the highlight
is right at the start or end of a clip, resulting in loss of
about 0.7 seconds of the highlight. So this commit makes the
hat function quite shallow, thereby reducing the highlight
loss in that case to about 0.2 seconds  (which is the descriptor
sampling interval) in those edge cases.