0023057: Saving to HD output crashed
A follow up to commit r8276 which allows you to
make the buffer behaviour during saving identical to
that during preview. This is useful to test some cases
with video which appear to succeed during preview
but hang or crash during saving.

Create the file c:\muveeDebug\DisableVideoBufferReduction.txt
to force the same buffering behaviour for both preview
and save runs.