#0024928: Add "Film Color" option to all built-in styles
The saturation slider had a range of 0 to 4. 0 is monochrome and 4 is super saturated. 1 is normal video. With this arrangement, the slider would have to be on the quarter mark by default to show normal video. However we want the slider to be in the middle position by default. So we made the slider range go from 0 to 2. But when we feed that value to the Cg program. we square the value.