0025072: Cg issue on Livermore
Unified CGcontext creation. Previously one context was
created for every compiled shader, which is kind of wasteful,
given that we don't know how much resource a CGcontext takes
up. So I've modified the code to use one CGcontext for
all CG shaders.

Checked this code and consequences with #rogbeer (Chetan).