#0024928: Add "Film Color" option to all built-in styles ColorModes have been added to all default style. We were sending in 5 parameters to the Cg programs. They were:
1. saturation (of UCG)
2. strength (of color effect)
3. TintR
4. TintG
5. TintB

It was all good and we were a bunch of really happy chipmunks... until muFAT started failing on DeepThought and Benjy.
It turns out they can only accept a maximum of 4 parameters.

So what we did was to allow vectors of data to be sent over instead of scalers. In other words, instead of sending in TintR, TintG and TintB. we now send one variable called Tint. Tint can be broken down into Tint.r, Tint.g and Tint.b

Doing that involved:

1. Parsing the parameter name string.
2. Looking for the '.' character
3. identifying if the letter after it was an 'r', a 'g', or a 'b'.
4. Storing it in a vector.
5. Sending the list of vectors to Cg.