Resolved #25449: OpenGL error, when length of title string exceeds a certain number of lines using Rolling text effect.
r12044 introduced this bug while making the TextOut
effect share textures from the pool. Fixed it by
making the pooled texture adapt in size to the
requirements of the given text string.

Tests: Tested the given case as well as with the typewriter
effect which dynamically changes the text string.