Resolved: #0025462: Encounter error code 40 during uploading muvee to shwup
It's because that these computer never play mp4 file before and mp4 is not in the registry, where ATL used to get content type from. So content type becomes application/octet-stream. Server side will take care of this, and reveal will also take care of this, map it to video/mp4 if extension if .mp4.