Resolved #25437: Black frames seen in preview & saved muvee when number of pictures is too large to fit to muvee duration. 0025072: Cg issue on Livermore
The problem of black frames was introduced in r11820 (see ~73820).
The problematic code was that the shader did not get compiled
in time when the shader did not have dynamic parameters.
If the shader had at least one dynamic parameter, it will
have worked fine.

The fix is to compile the shader at the point of use
if the shader state has changed. Even though the Compile()
call is placed in the frame loop, it won't be compiled
on every frame - just once and whenever the shader code
state has changed - i.e. whenever the actual code path
that the shader should take changes.