0025568: Application goes into an unresponsive state during construction with any style.
Modified CgFragmentShader class to throw an exception when
shader profile initialization fails. Also modified the
MESPreview::MESGetpreviewPos method to return S_FALSE when
the renderer is dead, which can happen if initialization
fails before preview starts.

As it stands, the app will now not be stuck in "Buffering..."
if initialization of rendering fails for some reason and will go
into the "stopped" state. Although that is not ideal and the app
should really be putting up a dialog box, it is much better than
being stuck in "Buffering.." with the whole UI disabled.

Next TODO: UI to respond to the S_FALSE returned by the
GetPreviewPos function and put up an appropriate dialog box.