Resolved #25665: Applying exclusion and split video, application crashes upon making muvee
Due to the change in sdkruntime to filter the list of includes per
video clip instead of per video file (~72117, r10832, #24681), the check in the constructor
to load and cache slice data based on the contents of the include
list of a video file breaks.

The fix is simple - to update the caching logic to check for slice
data again the first time a non-empty include list is returned for
a video clip. This explains Ernest's observation that excluding second
and other clips doesn't cause a crash - because the slice data is
loaded when examining the first clip.

Risk factor of this commit is low, so QA might want to consider
including this fix in Anderson_BR1_branch for the first launch.