Resolved #25670: Application crashes while switching aspect ratio or toggle between normal and FSP modes
The release cycle for the pooled textures was being done at the start
of render instead of the end of the render. That is usually ok, but if
you use captions and switch between 4:3 and 16:9 it triggers a condition
in which a pooled texture is first released in the resource manager's
Load() call before the render operation begins.

This fix places the recycling of pooled textures at the end of a frame's
render.