RESOLVED: #0025699: Rolling text is truncated in saved muvee. #yeecheng
When you add a new line to your title or credits text, it is denoted by \r\n.
However when that text is saved into a project file, a new line is denoted by \n.
Textbase was increasing its new line count by 4 every time a new line was found. 
So if you have two new lines that are close to one another, the increment of 4 meant that we were skipping some new line occurrences. I now increment in string search count by 2 instead of 4.