Resolved #25707: Manual duration of the muvee clip shorter than the duration of the first audio file, audio can be heard in the "credit" portion
Earlier, the following "if" condition also checked for
   srcIndex >= m_totalNumMusicFiles-1
which doesn't make sense in the case that the muvee duration is
shorter than that given by all the music files put together.
The m_muveeDuration check needs to be performed in the case srcIndex is
less than m_totalNumMusicFiles-1 as well. Removing that extra
condition fixes this bug.