0013043: muSE-X related changes
Introducing "ShaderOps" - for inserting shader settings
into the rendering pipeline in a way that lets them co-exist
with other ImageOps such as TextureSubset, and lets them
travel along with objects that are postponed for drawing
for the purpose of z-sorting followed by alpha composition.

The "Cg" effect is the only effect that current works using the
ShaderOp mechanism and therefore serves as an example about how to
modify other "shader" effects to behave like Cg. With this modification,
the "Cg" effect no longer needs its "GlobalMode" parameter and will
automatically do the right thing irrespective of whether it is
used for direct drawing or delayed alpha-composition.

Similar to ImageOps, only one ShaderOp can be in effect for any segment
of the scene graph.

Smoketests with several styles show that the Cg effect works fine.

TODO: Many of our current styles jump through hoops to get the shaders
to interoperate well with other effects and media in the scene. For example,
the CubeTwist style uses an explicit no-op shader on captions. Such work
arounds are no longer needed with the Cg effect now working using the
ShaderOp mechanism.