armCleaner.exe is a console program used for restoring the armadillo protection to its initial locked state
The xmls are the configs for the registry key locations and temp files created by armadillo protection.
To use, pass in the <application>.xml to armCleaner.exe. E.g. ">armCleaner.exe Anderson.xml"

Note that when building the armCleaner.exe, we need to arm the exe itself with ArmCleaner_09122009.arm file.
This is to enable proper retrieval of FingerprintInfo via the virtual copy of ArmAccess.dll that is built into the armed executable.