0025758: Santa and Hexplode's rapid overlays flicker badly in saved muvees
Chetan noticed another bug in Santa somewhat related to this
issue. If you make a muvee with 12 photos in the santa style, wait
in preview until the santa slides across the screen in the sleigh
and immediately hit "stop" followed by "play", and wait until that
point shows up, you'll notice that the texture is corrupted.

This commit fixes that issue. The problem is that upon hitting
stop, the recyclable resources such as rapid overlay textures
are not released. In addition to that, RapidOverlay's sequence
state also contributes to an invalid texture being returned.