0013043: muSE-X related changes
Changed the default value of preferred-transition-duration so that
it will be 0.0 whenever muvee-transition is undefined. This ensures
that the cuts in the resultant muvee are genuine cuts, segment level
effects will run from start to finish instead of, say, from 0.2 to 0.8
of progress ... and highlighted segments will not appear trimmed.

#tengchwan checked that none of the shipping styles will be affected.