0013043: muSE-X related changes
EXPERIMENTAL SUPPORT FOR MULTI-STYLE MUVEES

When the file c:\muveeDebug\concat.txt is present, every muvee you
make for preview will be appended to a running concatenated uber muvee.
It is the concatenated muvee that will play in the preview window.
When you remove this file, regular muvee making resumes, but the
concatenated muvee continues to reside in memory. If you add the file
again, you can resume concatenation from where you left off.
Quitting Reveal will destroy the concatenation.

When c:\muveeDebug\concat.txt is present and you save a muvee, the
last state of the concatenated muvee will be saved. Note that the
concatenated muvee will *not* be remade.

STABILITY: Not very stable. In particular, do not use Kinetic.
Some styles don't work in concatenation mode. This is just
a proof-of-concept and to thrash out architectural issues with
such concatenation.