0019759: muFAT crashes and assert failures - The Collection
Fixed a muFAT crash with r12810 in the constructor suite. The muFAT runs
all pass NULL pointers for the logo path when no logo is set,
which the code doesn't anticipate. The application passes an empty
string in those cases. Therefore muFAT crashes where the app chugs along.

This fix anticipates a null pointer for the logo path. On my machine,
a crashing muFAT run went through fine after this fix.

Note: 12810 is being used to demo in CES 2010. This muFAT crash case
should have no impact on the demo because the app never passes a null
pointer for the logo path.