#0025881: Add support for Nvidia 3D Vision to mES
Major reorganisation of the stereo vision code.

* The original back_left and back_right scene render in showTime_render_fullscreen has been completely deleted. All we do in that file is to create a stereo context.

* There was a lot of code inside ShowTime_RenderEnvBase to make sure that orthographic projection was disabled if we were using stereo vision. This is no longer needed and has been deleted.

* The perspective effect has been modified to accept parameters "StereoOffset" and "EnableStereo". Essentially the stereo vision code has been moved in the Perspective effect. This greatly simplifies the code and allows non-3D styles to co-exist with 3D styles on a 3D capable machine.

So lets dwell a bit on the rationale behind this whole thing.

We currently have a Cube3D style that looks uber cool with the shutter glasses. The problem is that all other styles end up looking like crap. So we need a system when non-3D styles can co-exist peacefully with 3D styles and maintain Reveal Harmony.

This is achieved by the "EnableStereo" flag that has been added to the Perspective effect. The old Perspestive effect was incompatible with stero. So we now branch off in that effect so we can render to the two back buffers.

Also we can now parametize the field of view for stereo vision. This is very important when we want to fine-tune the 3D pop-out-of-the-screen effect for each individual style.

Finally since we disabled orthographic projections by default in stereo vision, all the titles and credits for 3D as well as non-3D styles were severely distorted. This commit also fixed that.