#0019759: muFAT crashes and assert failures - The Collection
mufat runs on some machines have been crashing sporadically for a while now. We suspect Cg has something to do with it. So what I have done right now is to add 10 MVLOGs inside CgFragmentShader.cpp 

This is a temporary mesasure and will only serve to pinpoint where the crash is happening if it happens inside the Cg context.

After this test is complete, the MVLOGs will be deleted by next week.