#0026010: Logo enhancements
We need to minimize the use of the "render" effect.

Previously, we had one for the logo and one for the copyright notice.

The problem with 'render' effect is that it does a render-to-texture. This is super slow was data is juggled from the GPU to the CPU and back to the GPU.

Ok I've now optimized the code to use the 'render' effect only once.