#0026010: Logo enhancements
We were wrongly checking if the caption is *empty* before trying to display it. This condition has been inverted.
If the font string is empty, we do not set it. We instead use the default font of textout effect.
I've temporarily commented out HorizAlign and VertAlign parameters. They are not set by the UI.