#0026010: Logo enhancements
Stacked the perspective effect on top of the scale and translate effects for the copyright colorclip.
As it stands, we do not really require this effect. But muFAT is testing colorclips of variable lengths, making the perspective effect a requirement. 
So in the future, if we decide to support variable lengths of colorclips, no code change will be required within stylebase.