0026097: Hang in constructor when faced with insufficient media
The constructor hangs in a loop where it tries to grow the duration
of the muvee towards the desired duration exponentially closer. To
prevent that, it looks like all I needed to do is to make it shoot
for slightly (1/2 second) beyond the desired muvee duration.

Not marking this bug as resolved until I get a few mufat runs
to double check it.