0025904: S00548_Wedding1 Resolved #0026139: Encounter OpenGL Error in make muvee using Quiet Romance style
The problem is that tex2Dbias was used incorrectly. According to Cg documentation, the third argument is supposed to be an integer offset value, not float as we used it. Moreover text2Dbias function with 3 arguments is unsupported in profiles older than gp4. This caused the crash for some machines. A safer method is to use the function variant with 2 arguments. Then LOD value can be passed as the last component of the second parameter to achieve the intended output.

This fix worked for WS-CHETAN2.