0026122: Crash while saving Kinetic muvee with mt2s input to WMV-HD
merge from trunk rev 13643

Cleaned up the protocol to communicate with VideoServer
and made it 8-byte aligned. I expected a speed difference
after the alignment, which didn't happen, but I'm committing
this new protocol because it is better documented and in
some ways cleaner than the earlier version.

For an overview of the protocol structures, see VideoServerProtocol.h