#0026194: General code cleanup
Deleting unused GL_Shader_api.

It was written by Shiva a few years ago when we tried to use GLSL in muvee Reveal. Since our OpenGL version support was 1.4, and GLSL support was 2.0, it was never used. Now it's time to delete it. :) 

If a few years from now, we decide to implement a GLSL compiler, we'll just build on the existing CG compiler that we have.