#0026194: General code cleanup
Ok I found out why we needed MVGL. It's not because of any variable or functions inside those two files.

Its because... there was a pragma comment library in the MVGL.c file that MESCore actually needs.

So I am not deleting MVGL once again and I have moved the pragma comment library inside the vcproj of MESCore.cpp

The library name is glew32mx.lib