@kumar
@major milestone in MAC port!

 mesFoundation is now working RESONABLY WELL!!!!

FBase.cpp has a new set of functions called FGlobalLock() FGlobalUnlock() that are a set of mutex to print MVLOGs. Failure to include those will result in live-locks.

FBase.h has a new class called MVScope. It is currently commented out and should be used to test of thread live-locks / dead-locks.

FMutex.cpp  and FThread.cpp have a whole bunch of these MVScopes declared to be track thread movement.

FObject.cpp: We seem to be having a problem with OSAtomicIncrement32 and OSAtomicDecrement32. We do not know why. Therefore for the time being, we have replaced it with a set of FGlobalLock/FGlobalUnlock

main.cpp has some minor debugging tweaks added.

MESFoundation.xcodeproj has had some vector flags enabled in Release mode.


MVDiagnostics has a really neat feature added. Instead of printing all the Debug files in real-time,  we can accumulate everything in a stack and then display it at the end of the program execution. This is done mainly to ensure proper thread synchronisation and is to be used as a debugging tool