#0026194: General code cleanup
After deleting it, adding it and re-deleting it... I was thinking it'd be fun to add it for one last time.

Ok I previously deleted MVGL thinking we don't need it. Then I had to add it again because MESCore wouldn't link. When i found it was due to a #pragma comment within MVGL, i moved glew32mx.lib to the MESCore.vcproj project settings and deleted MVGL again.

Now I need to add it back. 

The reason?

It turns out that if you compile MESCore as an exe file (which we hardly ever do), MVGL is used to create the window in which the openGL context is created.

Since I'm working on the MAC project and I need to step though the MES rendering loop, I'm actually using MESCore.exe

So there you go...