0026351: Preview 1min muvee using 200 input photos will always have jerky preview
Fixed MESFoundation's FData class w.r.t. its IStream
implementation. This can be used to cache compressed
images in memory in mES. (Though such a feature is not
enabled at the moment, I did try it out in the context
of proxies for photos and noticed a good performance
boost. So I'm keeping this change.)