0026351: Preview 1min muvee using 200 input photos will always have jerky preview
Implemented a prototype of support for proxy files for photos
so that large photos in the muvee don't cause the preview to
stutter badly at high speed. Tested this with Kinetic and
seems to perform rather well at 5 photos per second on the
average.

To test out this behaviour, create c:\muveeDebug\EnablePhotoProxies.txt
Proxy files are created in
  C:\Users\USERNAME\AppData\Roaming\muvee Technologies\photoProxies\