Fixed bug with number of inputs for these two effects.
This raises an interesting question. The numInputs for these two effects was zero and thus crashed when we tried to use them. Correctly setting the inputs to 1 made them work fine. Now the question is, how come they worked fine on a PC all this time. I've double-checked and indeed on a PC, the numInputs is set to zero. humm, what a strange world.