#19759: muFAT crashes and assert failures - The Collection
Fixed muFAT crash when regressing issue #25026 on 14 June 2010
with r15148 on the trunk. The crash happened in the constructor
as a result of changes to fix issue #26036 "Highlighted segment
is not included in the muvee production".

I've verified that the muFAT run 0025026.run now passes on my
machine with this fix when it was failing at the same point
as the other machines prior to this fix. I've also verified that
the constructed muvee with 0025026/fire.rvl is correct, meaning
it includes all the highlighted material without fail and none
of the excluded material.