#26400: Features for "Telling a Story" initiative
Fixed the cube-twist crash mentioned in ~80018.
The number of nested levels in Cube Twist was blowing
the model-view stack, which on my machine has a max
depth of 32. Thankfully, I found that the SRotate effect
was unnecessarily using one level of nesting and
there were many levels of SRotate in the scene evaluation.
So cutting out that one extra level brought the
scene complexity to within limits (well within limits
I think).

I've also added logs at the start of preview for information
about the various limits of the driver+card combo.