#26400: Features for "Telling a Story" initiative
"Squares" style had a bug where the captions wouldn't
display whenever the exposed segment duration is < 0.5.
This fix is primarily to address that case.

In addition, it moves the Z plane of the fade to white
transition's white image forward to 0.1 instead of 0.001
so that the two-line text stuff doesn't appear in front
of the transition. This doesn't affect normal behaviour for
this style.