#26400: Features for "Telling a Story" initiative
Modified styles which use some form of fade-to-color
transition by presenting either an image or a ColorQuad
atop both A and B segments so that the image or ColorQuad
is placed at a z offset of 0.1 instead of small value such
as 0.001. This is to help the two-line-text feature.

The affected styles were all tested by forcing the
fade-to-color transitions to be used and were found to
be ok after this change.

The affected styles are -

1. S00510_SportHex (Hexplode)
2. S00523_Basic5 (Reminisce)
3. S00524_Basic6 (Showcase)
4. S00537_Nativity (Starry Night)
5. S00538_LifeStory (Life Story)
6. S00539_Tech (Techno Buzz)
7. S00540_Splats (Splatz)
8. S00548_Wedding1 (Quiet Romance)
9. S00549_Wedding2 (Kiss)