I have added the color attribute to text. So we can set the font and the color now.
Most importantly, I have manage to make the text fit the whole screen. For the longest time ever (more than a week), i've been trying to figure out when the text only occupied a square texture with coordinates -1 to 1 instead of -1.33 to 1.33. Well it turns out I had to scale the texture to the aspect ratio. All this while I was debugging texture generation code and reading about CGImage. Awesome!

Anyways we're on track for a full-fledged textout feature on the MAC.