ImageHelper class used a lot of  "malloc" and free to request for memory space. The is C syntax. Within mES, we request for memory using "new" and "delete".
I have decided to standardize everything with "new" and "delete"  because muvee Reveal is written in C++.

One by-product of this change is that I have to change the buffers from void* to Byte*.

I printed a "sizeof(Byte)" and confirmed that it was 1 byte. Do note that Byte and BYTE are different datatypes of the same size.