#0026912: [AndersonBasic] Encounter OpenGL Error after applying a new colour for T/C backgound
The check for texture generation success has been made silent. We will not throw an assert if it fails. BTW only BR2 has this code. The trunk will not have this because it is a nice check to have. But because of the special circumstances of BR2, we will only print it on the console instead of showing a message box.