Image helper used to get pointers to  Byte buffers and then create bitmap contexts with them. It was working fine but there was always the problem of deciding whose responsibility it was to delete the buffer when we were done with it. Sometimes the GLResourceManager deleted the buffer, on other times TextOUT deleted the buffer and finally ImageHelper itself got to delete buffers in some circumstances. Basically I poorly designed the class.
Ok now what i'll do it give classes ownership of buffers used in their methods. A class can get a pointer to a buffer in another class. But it makes a local copy of the buffer and then starts working with it. Then it becomes very clear who owns what.