Resolved #26931: Title-to-body and body-to-credits transitions overlap with normal body transitions
The issue was surprisingly easy to fix. I'd forgotten that though the
title and credits were constructed before the body of the muvee, the
title->body and body->credits transitions are constructed *after* the
body of the muvee. This puts us in the sweet position of deciding those
transition durations before creating the transitions themselves.

Problem solved.

NOTE: The title->body transition and body->credits transition are limited
to 40% of the duration of the segment that they overlap. This fraction is
taken from InputInfo::maxTransitionToSegDurRatio which is initialized to
0.4f in ShotInfo.cpp