#0026957: mES needs to be ported to the MAC
This is an interim commit for text shadows. We're almost there. Essentially I've currently hacked Ultra_Plain to include blurred layers of the negative of the text string beneath the user caption to achieve a new shadow effect. I'm not committing the ultra plain scum code because i will convert the muse code to C++ code and place it inside the TextOut effect, hence the interim nature of it all.

I've renamed many variables inside textout to different the text variables from the shadow variables.

I've fixed a bug whereby the textcolor had a range of 0 to 255. Opengl is expecting a range of 0.0 to 1.0