#0026957: mES needs to be ported to the MAC
Hortizontal alignment is working fine.
It looks like vertical alignment is not supported on Core text. We need to go one level up to Cocoa to get that feature. This is a luxury we can't afford. 

I might have to manually implement vertical alignment inside textOut. I'll cross that bridge on monday.