0026400: Features for "Telling a Story" initiative
BUGFIX: The "distance field text renderer" had a texture leak
because the SText class would hold onto the character textures
until well after the rendering context is destroyed. It was doing
this because it was taking an additional reference to the relevant
FImageTexture objects while their lifetimes were supposed to be
managed by the ResourceManager object.

This bug fix simply changes the FImageTexture references in SText
to raw pointers instead of storing them in FArray containers which
reference count their objects.