0026243: mES performance
Added an "audio buffer pool" mechanism for WAV file reader
to use a number of buffers that is constant relative to the
total number of AudioFile objects created for the purpose of
rendering an mES timeline. The number of buffers now depends
only on the maximum number of simultaneously playing audio
snippets. Note: It is not *equal* to that number, because audio
is rendered in chunks of 10 seconds (I think), but still it will
be relatively constant and independent of the total number of
audio clips.