#0025904: S00548_Wedding1
The Cg code for the gradient fade transition has some issues with it. Essentially we have a symmetric progress for both forward and reverve gradient fades. So when both input A and input B have a variable alpha, the common region will look dark. That's because we can see what's behind both pictures and that happens to be a black screen.

There's no easy way to fix this bug. For now, I am adding some code that is commented out by default. If it really annoys some people and they are only using video files with the new text effect, then they can open the scum file and uncomment some code to make the dark regions disappear.

At a later time before we do a refresh, I will revisit this style again and implement and asynchronous progress for input A and B.