0026400: Features for "Telling a Story" initiative
Jiayi found out that "bold" wasn't working (thanks Jiayi!).

I found that the font weight being passed in was 700
(800 is considered to be bold in general and 400 normal)
and I was checking for a weight value > 700 to enable bolding.
I just reduced the threshold to 600 - which is midway between
400 and 800 and should be a reliable threshold for deciding
bolding.

Bold now works.