0027310: Boost sountrack during selected clips and duck during VO clips
Modified Gerry's implementation of sync sound boosting during highlights
to boost any segment containing media from a clip whose sync sound the
"user" has requested to be boosted.

STATUS: You can indicate which clips to boost using c:\muveeDebug\SyncBoost.txt.
That is a plain text file containing a sequence of unique IDs of the clips
whose sync sound is desired to be boosted. This is a STOP GAP MEASURE until
the UI setting is propagated to the SDK.

TESTING: I've done some cursory testing on it and it seems to work fine.
Only one issue was of concern to me. Gerry had marked a point in code where
he was expecting the vector of highlight intervals to not overlap with
each other. However, when I tried it, it seems to be working fine even
when the boosted clips overlap. Anyway, I'm just leaving an MVFAILED
in there so that we return to that topic before this feature is
finalized.