0027310: Boost sountrack during selected clips and duck during VO clips
Apparently I was mistaken that the boost api wasn't implemented
at the SDK level. This commit therefore removes the SyncBoost.txt
hack since the app ui can be used now. The code is left in there
with a #if 0 so that you can re-enable it if you want.

I've also cleaned up the cause for why I was misled. I've moved
the sync sound boost info functions into its own interface named
IMVSyncSoundBoostInfo. Only the video source implements this
interface, therefore there is no need for a default implementation
for other source types.

One consequence of this change is that the latest "version number"
of the media sources is IMVMediaSource8 and not IMVMediaSource9.