0027139: New source type for TAS
Added a new style interface to return version number of the style. This is done by checking the FileVersion field of the version.dll in the style's resource folder. If this file does not exist, for example on development machines, it will return null.

Follow-up to commits 16396 and 16409. Checked in a better method of checking styles with text animation support (though not entirely foolproof) by extracting major version number from version.dll.