0026351: [AndersonBasic] Preview 1min muvee using 200 input photos will always have jerky preview
Temporarily switching back to mES's own photo proxy mechanism.
The proxies generated by the analyzer cause stutter during
preview 'cos their dimensions are not power of two. The
proxies generated by mES are power of two and so result in
smoother preview.

I don't know if this can be resolved in such a way that the
same proxy file can be used by both the app and mES.